Model API response from Polygon exchange rate API

+++ src/main/java/com/truebubo/maniflow/Currency/CurrencyConverter.java
@@ -3,5 +3,11 @@ package com.truebubo.maniflow.Currency;
 import java.math.BigDecimal;
 
 public interface CurrencyConverter {
-    BigDecimal convert(CurrencyDesignation from, CurrencyDesignation to);
+    /**
+     * How much of the {to currency} can we buy for 1 {from currency}
+     * @param from Currency we sell
+     * @param to Currency we buy
+     * @return exchange rate
+     */
+    Double convert(CurrencyDesignation from, CurrencyDesignation to);
 }
